+++ src/components/NostoProductCard.ts
@@ -1,6 +1,7 @@
 import { customElement } from "./decorators"
 import { evaluate } from "@/services/templating"
 import { assertRequired } from "@/utils"
+import { NostoElement } from "./NostoElement"
 
 /**
  * A custom element that renders a product card based on Nosto recommendation data.
@@ -36,7 +37,7 @@ import { assertRequired } from "@/utils"
  * ```
  */
 @customElement("nosto-product-card")
-export class NostoProductCard extends HTMLElement {
+export class NostoProductCard extends NostoElement {
   static attributes = {
     template: String
   }